Draw Mii face with split depth rather than with bald head
Requesting a split depth render results in getting two renders in one
image, the top half having the back half of the Mii head (usually the
hair) and then the bottom half having the front half.

This is needed for the Mii editor because it's drawing the body
manually, and you want some elements of the head (hair) to be behind the
body and the rest to be in front.

Previously, it solved this problem by loading a copy of the Mii head
that's bald, but this along with other stuff added up to four Mii Studio
requests on every render - now it's just one.

NOTE!!!!: This actually doesn't handle importing mii-js from the new
TypeScript project, so as of this commit it WILL NOT work - installing
the new repo, which this doesn't even reference, doesn't produce
compiled JS from the TS source, so it will simply not import.