feat: add terms
These terms are almost entirely a mashup of generated/pre-made documents from both https://termly.io/ and https://www.termsfeed.com/, which were then merged together and modified to the best of our ability to fit our specific needs. However we are not legal experts, and there is quite a lot of legalese here, and possibly broad policies with regards to collection/retention/sharing which we do NOT necessarily want/need. Our goal is to collect as LITTLE information about users as possible, and to never sell or directly share that data with any 3rd parties besides those required for operation, such as Cloudflare and Stripe. I have done my best to remove all the broad policies that would allow us to sell user data, as we do NOT want to do that, but I am no legal expert and may have missed things. In those cases, we are more than happy to revise/remove the problematic areas to clarify our goals/intentions